Create MemoizedFibonacci_CacheClass
The goal of this exercise was to improve the performance of a recursive Fibonacci function by caching the calculated elements of the secuence, thus avoiding many redundant calculations. Not knowing any better I implemented a singleton class to manage the cache independently from the Fibonacci function calls, which did the trick.

From other user's answers I learned that there is a more "Pythonic" way to deal with this situtation, which seems to involve what are called "decorators": I should look into those!