Create Product_of_consecutive_Fib_numbers
In Python you can declare multiples variables in a single line. In this exercise it turns out to be particularly useful, since at that line there is only access to the previous values of the variables. If you updated the variables in different lines you would need an auxiliary variable to save the previous value of "a" before it is updated, so you can use it to update "b".